fix done app system

+++ application/controllers/dashboard.php
@@ -1,14 +1,19 @@
 <?php 
 class Dashboard extends CI_Controller{
+	public function __construct(){
+		parent::__construct();
 
-	public function index()
-	{
-		$data['barang'] = $this->model_barang->tampil_data()->result();
-		$this->load->view('templates/header');
-		$this->load->view('templates/sidebar');
-		$this->load->view('dashboard',$data);
-		$this->load->view('templates/footer');
+		if($this->session->userdata('role_id') != '2'){
+			$this->session->set_flashdata('pesan','<div class="alert alert-danger alert-dismissible fade show" 				role="alert">
+  								silahkan login !!!
+ 								 <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+    								<span aria-hidden="true">&times;</span>
+  								</button></div>');
+			redirect('auth/login');
+		}
 	}
+
+
 	public	function	tambah_ke_keranjang($id)
 	{
 		 	$barang	= $this->model_barang->find($id);
@@ -21,7 +26,7 @@ class Dashboard extends CI_Controller{
 		 		);
 
 		 $this->cart->insert($data);
-		 redirect('dashboard');
+		 redirect('welcome');
 	}
 	public	function detail_keranjang()
 	{
@@ -33,7 +38,7 @@ class Dashboard extends CI_Controller{
 	public function hapus_keranjang()
 	{
 		$this->cart->destroy();
-		redirect('dashboard/index'); 
+		redirect('welcome'); 
 	}
 	public function pembayaran()
 	{
@@ -44,10 +49,23 @@ class Dashboard extends CI_Controller{
 	}
 	public function proses_pesanan()
 	{
+		$is_processed = $this->model_invoice->index();
+		if($is_processed){
 		$this->cart->destroy();
 		$this->load->view('templates/header');
 		$this->load->view('templates/sidebar');
 		$this->load->view('proses_pesanan');
 		$this->load->view('templates/footer');	
+		} else{
+			echo "maaf, pesanan gagal di simpan!!!!!!!!!!";
+		}
+	}
+	public function detail($id_brg)
+	{
+		$data['barang'] = $this->model_barang->detail_brg($id_brg);
+		$this->load->view('templates/header');
+		$this->load->view('templates/sidebar');
+		$this->load->view('detail_barang', $data);
+		$this->load->view('templates/footer');
 	}
 }
